ai-engine/requirements.txt
web3
requests
python-dotenv
numpy
pandas
scikit-learn
fastapi
uvicorn
path
sys
json
os